Clean up identity mock code  - Move MockCaClient to caclient.rs  - Get rid of the "Mock" prefix when inside the mock module  - Unexport identity submodules, they are an implementation detail  - Add a mock::new_secret_manager function to construct SecretManager<MockCaClient>, having MockCaClient::new return a SecretManager is confusing
