+++ src/app.rs
@@ -132,7 +132,7 @@ pub async fn build_with_cert(
 
 pub async fn build(config: config::Config) -> anyhow::Result<Bound> {
     if config.fake_ca {
-        let cert_manager = identity::mock::MockCaClient::new(Duration::from_secs(86400));
+        let cert_manager = identity::mock::new_secret_manager(Duration::from_secs(86400));
         build_with_cert(config, cert_manager).await
     } else {
         let cert_manager = identity::SecretManager::new(config.clone())?;
